requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Injector==0.13.0
injector==0.18.4
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
mongoengine==0.23.1
pymongo==3.12.0
python-dotenv==0.19.1
Werkzeug==2.0.1
imgviz==1.2.2
matplotlib==3.3.1
moviepy==1.0.3
natsort==7.1.1
numpy==1.19.1
opencv-python==4.5.4.58
Pillow==7.0.0
scikit-learn==1.0
scipy==1.7.1
simplejson==3.16.0